+++ R/methods_MCMCglmm.R
@@ -34,6 +34,7 @@ model_parameters.MCMCglmm <- function(model,
                                       bf_prior = NULL,
                                       diagnostic = c("ESS", "Rhat"),
                                       priors = TRUE,
+                                      parameters = NULL,
                                       verbose = TRUE,
                                       ...) {
 
@@ -51,6 +52,7 @@ model_parameters.MCMCglmm <- function(model,
       bf_prior = bf_prior,
       diagnostic = diagnostic,
       priors = priors,
+      filter_parameters = parameters,
       verbose = verbose,
       ...
     )