+++ R/methods_bayesQR.R
@@ -10,6 +10,7 @@ model_parameters.bayesQR <- function(model,
                                      bf_prior = NULL,
                                      diagnostic = c("ESS", "Rhat"),
                                      priors = TRUE,
+                                     parameters = NULL,
                                      verbose = TRUE,
                                      ...) {
 
@@ -27,6 +28,7 @@ model_parameters.bayesQR <- function(model,
     bf_prior = bf_prior,
     diagnostic = diagnostic,
     priors = priors,
+    filter_parameters = parameters,
     verbose = verbose,
     ...
   )