+++ R/methods_brms.R
@@ -18,6 +18,7 @@ model_parameters.brmsfit <- function(model,
                                      exponentiate = FALSE,
                                      standardize = NULL,
                                      group_level = FALSE,
+                                     parameters = NULL,
                                      verbose = TRUE,
                                      ...) {
   modelinfo <- insight::model_info(model)
@@ -38,6 +39,7 @@ model_parameters.brmsfit <- function(model,
       priors = priors,
       exponentiate = exponentiate,
       standardize = standardize,
+      filter_parameters = parameters,
       ...
     )
   } else {
@@ -58,6 +60,7 @@ model_parameters.brmsfit <- function(model,
       effects = effects,
       component = component,
       standardize = standardize,
+      filter_parameters = parameters,
       verbose = verbose,
       ...
     )
@@ -106,6 +109,7 @@ model_parameters.brmsfit <- function(model,
                                         priors = FALSE,
                                         exponentiate = FALSE,
                                         standardize = NULL,
+                                        filter_parameters = NULL,
                                         verbose = TRUE,
                                         ...) {
 
@@ -177,6 +181,9 @@ model_parameters.brmsfit <- function(model,
   first_cols <- c(1:ci_column, weight_column)
   params <- params[, c(first_cols, seq_len(ncol(params))[-first_cols])]
 
+  if (!is.null(filter_parameters)) {
+    parameters <- .filter_parameters(parameters, filter_parameters)
+  }
 
   # add attributes
   attr(params, "tau") <- params_tau