+++ R/methods_cplm.R
@@ -36,6 +36,7 @@ model_parameters.zcpglm <- function(model,
                                     exponentiate = FALSE,
                                     robust = FALSE,
                                     p_adjust = NULL,
+                                    parameters = NULL,
                                     verbose = TRUE,
                                     ...) {
   component <- match.arg(component)
@@ -48,26 +49,27 @@ model_parameters.zcpglm <- function(model,
 
   # Processing
   if (bootstrap) {
-    parameters <- bootstrap_parameters(model, iterations = iterations, ci = ci, ...)
+    params <- bootstrap_parameters(model, iterations = iterations, ci = ci, ...)
   } else {
-    parameters <- .extract_parameters_generic(
+    params <- .extract_parameters_generic(
       model,
       ci = ci,
       component = component,
       standardize = standardize,
       robust = robust,
       p_adjust = p_adjust,
+      filter_parameters = parameters,
       ...
     )
   }
 
 
   if (isTRUE(exponentiate) || identical(exponentiate, "nongaussian")) {
-    parameters <- .exponentiate_parameters(parameters, model, exponentiate)
+    params <- .exponentiate_parameters(params, model, exponentiate)
   }
 
-  parameters <- .add_model_parameters_attributes(
-    parameters,
+  params <- .add_model_parameters_attributes(
+    params,
     model,
     ci,
     exponentiate,
@@ -75,10 +77,10 @@ model_parameters.zcpglm <- function(model,
     verbose = verbose,
     ...
   )
-  attr(parameters, "object_name") <- deparse(substitute(model), width.cutoff = 500)
-  class(parameters) <- c("parameters_model", "see_parameters_model", class(parameters))
+  attr(params, "object_name") <- deparse(substitute(model), width.cutoff = 500)
+  class(params) <- c("parameters_model", "see_parameters_model", class(params))
 
-  parameters
+  params
 }