+++ R/methods_lavaan.R
@@ -39,6 +39,13 @@
 #'   model_parameters(model)
 #'   model_parameters(model, standardize = TRUE)
 #'
+#'   # filter parameters
+#'   model_parameters(
+#'     model,
+#'     parameters = list(To = "^(?!visual)",
+#'                       From = "^(?!(x7|x8))")
+#'   )
+#'
 #'   # Structural Equation Model (SEM) ------------
 #'
 #'   structure <- "
@@ -75,12 +82,14 @@ model_parameters.lavaan <- function(model,
                                     ci = 0.95,
                                     standardize = FALSE,
                                     component = c("regression", "correlation", "loading", "defined"),
+                                    parameters = NULL,
                                     verbose = TRUE,
                                     ...) {
 
   params <- .extract_parameters_lavaan(model,
                                        ci = ci,
                                        standardize = standardize,
+                                       filter_parameters = parameters,
                                        verbose = verbose,
                                        ...
   )
@@ -112,6 +121,7 @@ model_parameters.blavaan <- function(model,
                                      diagnostic = c("ESS", "Rhat"),
                                      component = "all",
                                      standardize = NULL,
+                                     parameters = NULL,
                                      verbose = TRUE,
                                      ...) {
 
@@ -129,6 +139,7 @@ model_parameters.blavaan <- function(model,
     diagnostic = diagnostic,
     effects = "all",
     standardize = standardize,
+    filter_parameters = parameters,
     verbose = verbose,
     ...
   )
@@ -210,7 +221,6 @@ p_value.blavaan <- p_value.BFBayesFactor
 
 
 # print ---------------------------
-
 
 #' @importFrom insight export_table
 #' @export