+++ R/methods_lme4.R
@@ -127,6 +127,7 @@ model_parameters.merMod <- function(model,
                                     p_adjust = NULL,
                                     wb_component = TRUE,
                                     summary = FALSE,
+                                    parameters = NULL,
                                     verbose = TRUE,
                                     ...) {
 
@@ -170,6 +171,7 @@ model_parameters.merMod <- function(model,
         standardize = standardize,
         p_adjust = p_adjust,
         wb_component = wb_component,
+        filter_parameters = parameters,
         verbose = verbose,
         ...
       )