+++ R/methods_rstanarm.R
@@ -55,6 +55,7 @@ model_parameters.stanreg <- function(model,
                                      exponentiate = FALSE,
                                      standardize = NULL,
                                      group_level = FALSE,
+                                     parameters = NULL,
                                      verbose = TRUE,
                                      ...) {
 
@@ -74,6 +75,7 @@ model_parameters.stanreg <- function(model,
     priors = priors,
     effects = effects,
     standardize = standardize,
+    filter_parameters = parameters,
     verbose = verbose,
     ...
   )
@@ -120,6 +122,7 @@ model_parameters.stanmvreg <- function(model,
                                        priors = TRUE,
                                        effects = "fixed",
                                        standardize = NULL,
+                                       parameters = NULL,
                                        verbose = TRUE,
                                        ...) {
 
@@ -139,6 +142,7 @@ model_parameters.stanmvreg <- function(model,
     priors = priors,
     effects = effects,
     standardize = standardize,
+    filter_parameters = parameters,
     verbose = verbose,
     ...
   )