+++ NEWS.md
@@ -4,6 +4,11 @@
 
 * Removed deprecated arguments.
 
+## Changes to functions
+
+* `model_parameters()` gains a `parameters` argument, which takes a regular
+  expression as string, to filter parameters from the returned data frame.
+
 ## Bug fixes
 
 * Arguments `digits`, `ci_digits` and `p_digits` were ignored for `print()`